requirements.txt
requests>=2.25.1
feedparser>=6.0.2
pandas>=1.3.0
numpy>=1.20.0
beautifulsoup4>=4.9.3
newspaper3k>=0.2.8
nltk>=3.6.2
transformers>=4.11.3
torch>=1.9.0